pyproject.toml
[project]
name = "backend_form"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "luansilva25", email = "luan250807@gmail.com"},
]
dependencies = [
    "django>=5.0.7",
    "pillow>=10.4.0",
    "djangorestframework>=3.15.2",
    "python-magic>=0.4.27",
    "django-cors-headers>=4.4.0",
    "psycopg2>=2.9.9",
    "psycopg2-binary>=2.9.9",
    "waitress>=3.0.0",
    "whitenoise[binary]>=6.7.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
